brave-search-mcp/requirements.txt
langchain
langchain_openai
autogen
requests
flask
python-dotenv
ipywidgets